requirements.txt
black==25.1.0
pygame==2.6.1
requests==2.32.3
python-crontab~=3.2.0
